requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
textual
python-dotenv
pydantic-settings
httpx
websockets
